Remove unused Tag Store
